add publisher.php

+++ publisher.php
@@ -2,48 +2,53 @@
     // require_once is the most common used to link the database connetion
     require_once 'components/db_connect.php';
 
-    $sql = "SELECT stock.*, 
+    if(isset($_GET['id']) && !empty($_GET["id"])){
+        $id_publisher = mysqli_real_escape_string($conn, $_GET['id']);
+        $sql = "SELECT stock.*, 
                     type.type AS typeName, 
                     author.first_name AS authorFirstName, 
                     author.last_name AS authorLastName, 
                     publisher.first_name AS publisherFirstName, 
                     publisher.last_name AS publisherLastName 
-            FROM `stock` 
-            INNER JOIN `type` ON stock.fk_type = type.id_type 
-            INNER JOIN `author` ON stock.fk_author = author.id_author 
-            INNER JOIN `publisher` ON stock.fk_publisher = publisher.id_publisher";
-    $result = mysqli_query($conn, $sql);    
-    $cards = "";
-    
-    if(mysqli_num_rows($result) > 0) {
-        while($row = mysqli_fetch_assoc($result)) {
-            $cards .= "
-            <div>
-                <div class='card'>
-                    <div id='img'>
-                        <img src='assets/$row[img]' class='card-img' alt=''>
-                    </div>
-                    <div class='card-body'>
-                        <h3>$row[title]</h3>
-                        <h5>$row[isbn]</h5>
-                        <p class='card-text'>Type: {$row["typeName"]}</p>
-                        <p class='card-text'>Author: {$row["authorFirstName"]} {$row["authorLastName"]}</p>
-                        <p class='card-text'>Publisher: <a href='publisher.php?id=$row[fk_publisher]'>{$row["publisherFirstName"]} {$row["publisherLastName"]}</a></p>
-                        <p class='card-text'>Publish Date: $row[publish_date]</p>
-                        <p class='card-text'>$row[description]</p>
-                        <a href='details.php?id=$row[id_stock]' class='btn btn-primary'>Details</a>
-                        <a href='update.php?id=$row[id_stock]' class='btn btn-warning'>Edit</a>
-                        <a href='delete.php?id=$row[id_stock]' class='btn btn-danger delete'>Delete</a>
+                FROM `stock` 
+                INNER JOIN `type` ON stock.fk_type = type.id_type 
+                INNER JOIN `author` ON stock.fk_author = author.id_author 
+                INNER JOIN `publisher` ON stock.fk_publisher = publisher.id_publisher
+                WHERE `stock`.`fk_publisher` = $id_publisher";
+        $result = mysqli_query($conn, $sql);    
+        $cards = "";
+        $publisher = "";
+
+        if(mysqli_num_rows($result) > 0) {
+            while($row = mysqli_fetch_assoc($result)) {
+                $cards .= "
+                <div>
+                    <div class='card'>
+                        <div id='img'>
+                            <img src='assets/$row[img]' class='card-img' alt=''>
+                        </div>
+                        <div class='card-body'>
+                            <h3>$row[title]</h3>
+                            <h5>$row[isbn]</h5>
+                            <p class='card-text'>Type: {$row["typeName"]}</p>
+                            <p class='card-text'>Author: {$row["authorFirstName"]} {$row["authorLastName"]}</p>
+                            <p class='card-text'>Publisher: <a href='publisher.php?id=$row[fk_publisher]'>{$row["publisherFirstName"]} {$row["publisherLastName"]}</a></p>
+                            <p class='card-text'>Publish Date: $row[publish_date]</p>
+                            <p class='card-text'>$row[description]</p>
+                            <a href='details.php?id=$row[id_stock]' class='btn btn-primary'>Details</a>
+                            <a href='update.php?id=$row[id_stock]' class='btn btn-warning'>Edit</a>
+                            <a href='delete.php?id=$row[id_stock]' class='btn btn-danger delete'>Delete</a>
+                        </div>
                     </div>
                 </div>
-            </div>
-            ";
-        }
-    } else {
+                ";
+                $publisher = "<p class='card-text'>: {$row["publisherFirstName"]} {$row["publisherLastName"]}</p>
+                ";
+            }
+        } else {
         $cards = "no data found";
+        }
     }
-        
-    mysqli_close($conn);
 ?>
 
 <!DOCTYPE html>
@@ -64,7 +69,8 @@
 
     <div class='stock'>
         <div class='headline'>
-            <h1>Books & more</h1>
+            <h1>Books & more from</h1>
+            <h1><?= $publisher ?></h1>
         </div>
         <div class='row row-cols-xl-5 row-cols-lg-4 row-cols-md-2 row-cols-sm-2 row-cols-xs-1' id='result'>
             <?= $cards ?>
